requirements.txt
google-generativeai>=0.3.0
rich>=13.0.0
click>=8.0.0
pyyaml>=6.0
aiofiles>=23.0.0
